+++ App.js
@@ -1,0 +1,18 @@
+/**
+ * Sample React Native App
+ * https://github.com/facebook/react-native
+ *
+ * @format
+ * @flow
+ */
+
+import React, {Component} from 'react';
+import Home from './src/scenes/home/Home'
+
+export default class App extends Component {
+  render() {
+    return (
+      <Home />
+    );
+  }
+}
