+++ index.js
@@ -3,7 +3,7 @@
  */
 
 import { AppRegistry } from 'react-native';
-import Home from './src/components/Home';
 import { name as appName } from './app.json';
+import App from './App'
 
-AppRegistry.registerComponent(appName, () => Home);
+AppRegistry.registerComponent(appName, () => App);
