+++ src/helpers/Colors.js
@@ -1,0 +1,8 @@
+const colors = {
+  white: '#ffffff',
+  clearBlue: '#1f86ff',
+  warmGrey: '#959595',
+  borderGrey: '#f0f0f0',
+};
+
+export default colors;
